Move Chunks protocol into Protocols to prepare for reuse in FDO2 driver
