+++ firmware/ventilator-controller-stm32/Core/Inc/Pufferfish/Driver/Serial/Backend/Frames.h
@@ -10,6 +10,7 @@
 #include <cstddef>
 #include <cstdint>
 
+#include "Pufferfish/Protocols/Chunks.h"
 #include "Pufferfish/Util/Vector.h"
 
 namespace Pufferfish::Driver::Serial::Backend {
@@ -19,39 +20,11 @@ struct FrameProps {
   static const size_t chunk_max_size = payload_max_size + 2;  // including delimiter
   using ChunkBuffer = Util::ByteVector<chunk_max_size>;
 
-  enum class InputStatus { ok = 0, output_ready, invalid_length };
-  enum class OutputStatus { ok = 0, waiting, invalid_length };
+  using InputStatus = Protocols::ChunkInputStatus;
+  using OutputStatus = Protocols::ChunkOutputStatus;
 };
 
-// Splits delimited chunks from a stream
-template <size_t buffer_size>
-class ChunkSplitter {
- public:
-  explicit ChunkSplitter(uint8_t delimiter = 0x00) : delimiter(delimiter) {}
-
-  // Call this until it returns available, then call output
-  FrameProps::InputStatus input(uint8_t new_byte);
-  FrameProps::OutputStatus output(Util::ByteVector<buffer_size> &output_buffer);
-
- private:
-  Util::ByteVector<buffer_size> buffer_;
-  const uint8_t delimiter;
-  FrameProps::InputStatus input_status_ = FrameProps::InputStatus::ok;
-};
-
-using FrameChunkSplitter = ChunkSplitter<FrameProps::chunk_max_size>;
-
-// Merges chunks into a stream by delimiting them
-class ChunkMerger {
- public:
-  explicit ChunkMerger(uint8_t delimiter = 0x00) : delimiter(delimiter) {}
-
-  template <size_t buffer_size>
-  FrameProps::OutputStatus transform(Util::ByteVector<buffer_size> &input_output_buffer) const;
-
- private:
-  const uint8_t delimiter = 0x00;
-};
+using FrameChunkSplitter = Protocols::ChunkSplitter<FrameProps::chunk_max_size>;
 
 // Decodes frames (length up to 255 bytes, excluding frame delimiter) with COBS
 class COBSDecoder {
@@ -97,7 +70,7 @@ class FrameSender {
 
  private:
   const COBSEncoder cobs_encoder = COBSEncoder();
-  const ChunkMerger chunk_merger = ChunkMerger();
+  const Protocols::ChunkMerger chunk_merger = Protocols::ChunkMerger();
 };
 
 }  // namespace Pufferfish::Driver::Serial::Backend
